Update ap (#35)
* Update index.html

* Update to index and further items functionality

* further developed the items functionality, now can add items of different object classes

* further development to the items module

* Update items_edit.html

* fixed donor repository so I could run

* updated naming convention for items in shop and sales

* Update to the item edits

* Further changes the the items and inventory module

* fixed the spelling mistake

* shop and cart formatting

* update to mysql query

* changes to user type visibility

* additional formatting

* update to index and error in donationscontroller

+++ src/main/java/com/AnimalLoversSociety/MyApplication/donations/DonationsController.java
@@ -27,7 +27,7 @@ public class DonationsController {
     @GetMapping(path="/donations/all/{id}")
     public Optional<Donations> show(@PathVariable String id) {
         long donationId = Long.parseLong(id);
-        return donationsRepo.findById((int) donationId);
+        return donationsRepo.findById(/*(int)*/ donationId);
     }
 
     //Delete method is not working for some reason
